SR-4572 Remove jinja2 dependency from benchmark test harness generation
