+++ benchmark/CMakeLists.txt
@@ -1,8 +1,7 @@
 # -*- mode: cmake -*-
 ################################################################################
 # WARNING: This file is automatically generated from templates and should not
-# be directly modified. Instead, make changes to
-# scripts/generate_harness/CMakeLists.txt_template and run
+# be directly modified. Instead, make changes to CMakeLists.text.gyb and run
 # scripts/generate_harness/generate_harness.py to regenerate this file.
 ################################################################################
