--- benchmark/scripts/generate_harness/main.swift_template
@@ -1,63 +1,0 @@
-//===--- main.swift -------------------------------------------------------===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
-// Licensed under Apache License v2.0 with Runtime Library Exception
-//
-// See https://swift.org/LICENSE.txt for license information
-// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
-//
-//===----------------------------------------------------------------------===//
-
-////////////////////////////////////////////////////////////////////////////////
-// WARNING: This file is automatically generated from templates and should not
-// be directly modified. Instead, make changes to
-// scripts/generate_harness/main.swift_template and run
-// scripts/generate_harness/generate_harness.py to regenerate this file.
-////////////////////////////////////////////////////////////////////////////////
-
-// This is just a driver for performance overview tests.
-import TestsUtils
-import DriverUtils
-{% for import in imports %}
-import {{ import }}
-{% endfor %}
-
-@inline(__always)
-private func addPrecommitTest(
-  _ name: String, _ function: @escaping (Int) -> ()
-) {
-  precommitTests[name] = function
-}
-@inline(__always)
-private func addStringTest(
-  _ name: String, _ function: @escaping (Int) -> ()
-) {
-  stringTests[name] = function
-}
-@inline(__always)
-private func addOtherTest(
-  _ name: String, _ function: @escaping (Int) -> ()
-) {
-  otherTests[name] = function
-}
-
-// The main test suite: precommit tests
-{% for run_func in run_funcs %}
-addPrecommitTest("{{ run_func[0] }}", run_{{ run_func[1] }})
-{% endfor %}
-
-// Other tests
-{% for test_name in other_tests %}
-addOtherTest("{{ test_name }}", run_{{ test_name }})
-{% endfor %}
-
-// String tests, an extended benchmark suite exercising finer-granularity
-// behavior of our Strings.
-{% for test_name in string_tests %}
-addStringTest("{{ test_name }}", run_{{ test_name }})
-{% endfor %}
-
-main()
-
