+++ benchmark/utils/main.swift
@@ -1,4 +1,4 @@
-//===--- main.swift -------------------------------------------------------===//
+//===--- main.swift -------------------------------------------*- swift -*-===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -11,9 +11,8 @@
 //===----------------------------------------------------------------------===//
 
 ////////////////////////////////////////////////////////////////////////////////
-// WARNING: This file is automatically generated from templates and should not
-// be directly modified. Instead, make changes to
-// scripts/generate_harness/main.swift_template and run
+// WARNING: This file is manually generated from .gyb template and should not
+// be directly modified. Instead, make changes to main.swift.gyb and run
 // scripts/generate_harness/generate_harness.py to regenerate this file.
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -412,7 +411,6 @@ addPrecommitTest("XorLoop", run_XorLoop)
 
 // Other tests
 addOtherTest("Ackermann", run_Ackermann)
-addOtherTest("Fibonacci", run_Fibonacci)
 addOtherTest("ExistentialTestArrayConditionalShift_ClassValueBuffer1", run_ExistentialTestArrayConditionalShift_ClassValueBuffer1)
 addOtherTest("ExistentialTestArrayConditionalShift_ClassValueBuffer2", run_ExistentialTestArrayConditionalShift_ClassValueBuffer2)
 addOtherTest("ExistentialTestArrayConditionalShift_ClassValueBuffer3", run_ExistentialTestArrayConditionalShift_ClassValueBuffer3)
@@ -512,20 +510,21 @@ addOtherTest("ExistentialTestTwoMethodCalls_IntValueBuffer1", run_ExistentialTes
 addOtherTest("ExistentialTestTwoMethodCalls_IntValueBuffer2", run_ExistentialTestTwoMethodCalls_IntValueBuffer2)
 addOtherTest("ExistentialTestTwoMethodCalls_IntValueBuffer3", run_ExistentialTestTwoMethodCalls_IntValueBuffer3)
 addOtherTest("ExistentialTestTwoMethodCalls_IntValueBuffer4", run_ExistentialTestTwoMethodCalls_IntValueBuffer4)
+addOtherTest("Fibonacci", run_Fibonacci)
 
 // String tests, an extended benchmark suite exercising finer-granularity
 // behavior of our Strings.
-addStringTest("StringWalkASCIIScalars", run_StringWalkASCIIScalars)
 addStringTest("StringWalkASCIICharacters", run_StringWalkASCIICharacters)
-addStringTest("StringWalkUnicodeScalars", run_StringWalkUnicodeScalars)
-addStringTest("StringWalkUnicodeCharacters", run_StringWalkUnicodeCharacters)
-addStringTest("StringWalkMixedScalars", run_StringWalkMixedScalars)
-addStringTest("StringWalkMixedCharacters", run_StringWalkMixedCharacters)
-addStringTest("StringWalkASCIIScalarsBackwards", run_StringWalkASCIIScalarsBackwards)
 addStringTest("StringWalkASCIICharactersBackwards", run_StringWalkASCIICharactersBackwards)
-addStringTest("StringWalkUnicodeScalarsBackwards", run_StringWalkUnicodeScalarsBackwards)
-addStringTest("StringWalkUnicodeCharactersBackwards", run_StringWalkUnicodeCharactersBackwards)
-addStringTest("StringWalkMixedScalarsBackwards", run_StringWalkMixedScalarsBackwards)
+addStringTest("StringWalkASCIIScalars", run_StringWalkASCIIScalars)
+addStringTest("StringWalkASCIIScalarsBackwards", run_StringWalkASCIIScalarsBackwards)
+addStringTest("StringWalkMixedCharacters", run_StringWalkMixedCharacters)
 addStringTest("StringWalkMixedCharactersBackwards", run_StringWalkMixedCharactersBackwards)
+addStringTest("StringWalkMixedScalars", run_StringWalkMixedScalars)
+addStringTest("StringWalkMixedScalarsBackwards", run_StringWalkMixedScalarsBackwards)
+addStringTest("StringWalkUnicodeCharacters", run_StringWalkUnicodeCharacters)
+addStringTest("StringWalkUnicodeCharactersBackwards", run_StringWalkUnicodeCharactersBackwards)
+addStringTest("StringWalkUnicodeScalars", run_StringWalkUnicodeScalars)
+addStringTest("StringWalkUnicodeScalarsBackwards", run_StringWalkUnicodeScalarsBackwards)
 
 main()
